requirements.txt
# Core Django Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
crispy-bootstrap5==0.7

# Advanced Database & Caching
psycopg2-binary==2.9.7
redis==5.0.1
django-redis==5.4.0
django-cacheops==8.0.1

# Background Tasks & Message Queue
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Advanced Authentication & Security
django-allauth==0.57.0
django-oauth-toolkit==2.3.0
django-ratelimit==4.1.0
django-axes==6.1.0
django-defender==0.9.0

# API & Serialization
django-filter==23.3
django-simple-history==3.4.0
djangorestframework-simplejwt==5.3.0
django-rest-auth==0.9.5

# Advanced Forms & UI
Pillow>=10.0.0
django-countries==7.5.1
django-phonenumber-field==7.3.0
django-address==0.2.3

# AI/ML & Data Processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
opencv-python==4.8.1.78
matplotlib==3.7.2
seaborn==0.12.2

# Real-time & WebSocket
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Advanced Monitoring & Logging
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-silk==5.0.3

# External APIs & Services
requests==2.31.0
python-decouple==3.8
twilio==8.10.0
googletrans==4.0.0rc1
python-dotenv==1.0.0

# Production & Deployment
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2

# Testing & Quality
pytest==7.4.2
pytest-django==4.5.2
coverage==7.3.2
factory-boy==3.3.0

# Development Tools
black==23.9.1
flake8==6.1.0
isort==5.12.0
pre-commit==3.5.0
